+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 .
 
+## [8.1.2] - 2021-11-15
+
+-   Uses supertokens-js-override from npm
+
 ## [8.1.1] - 2021-11-08
 
 ### Changes